Crate dstr

Modules

raw

Structs

CharOffsets
External iterator for a string's characters and their byte offsets. Use with the std::iter module.

CharRange
Struct that contains a char and the index of the first byte of the next char in a string. This can be used as a data structure for iterating over the UTF-8 bytes of a string.

CharSplits
An iterator over the substrings of a string, separated by sep.

CharSplitsN
An iterator over the substrings of a string, separated by sep, splitting at most count times.

Chars
Iterator for the char (representing Unicode Scalar Values) of a string

MatchIndices
An iterator over the start and end indices of the matches of a substring within a larger string

Str
Rust's string type

StrSplits
An iterator over the substrings of a string separated by a given search string

Utf16CodeUnits
External iterator for a string's UTF16 codeunits. Use with the std::iter module.

Traits

AsStr

Functions

eq_slice
Bytewise slice equality NOTE: This function is (ab)used in rustc::middle::trans::_match to compare &[u8] byte slices that are not necessarily valid UTF-8.

is_utf8
Determines if a vector of bytes contains valid UTF-8.

s

utf8_char_width
Given a first byte, determine how many bytes are in this UTF-8 character

Type Definitions

AnyLines
An iterator over the lines of a string, separated by either \n or (\r\n).

Bytes
External iterator for a string's bytes. Use with the std::iter module.