Struct dstr::Str

pub struct Str(_);

Rust's string type

Methods

impl Str

fn from_utf8<'a>(v: &'a [u8]) -> Option<&'a Str>

Converts a vector to a string slice without performing any allocations.

Once the slice has been validated as utf-8, it is transmuted in-place and returned as a '&str' instead of a '&[u8]'

Returns None if the slice is not utf-8.

fn contains(&self, needle: &Str) -> bool

Returns true if one string contains another

Arguments

needle - The string to look for

Example

assert!(s("bananas").contains(s("nana")));

fn contains_char(&self, needle: char) -> bool

Returns true if a string contains a char.

Arguments

needle - The char to look for

Example

assert!(s("hello").contains_char('e'));

fn chars(&self) -> Chars

An iterator over the characters of self. Note, this iterates over Unicode code-points, not Unicode graphemes.

Example

let v: Vec<char> = s("abc åäö").chars().collect();
assert_eq!(v, vec!['a', 'b', 'c', ' ', 'å', 'ä', 'ö']);

fn bytes(&self) -> Bytes

An iterator over the bytes of self

Example

let v: Vec<u8> = s("bors").bytes().collect();
assert_eq!(v, b"bors".to_vec());

fn char_indices(&self) -> CharOffsets

An iterator over the characters of self and their byte offsets.

fn split<Sep: CharEq>(&self, sep: Sep) -> CharSplits<Sep>

An iterator over substrings of self, separated by characters matched by sep.

Example

let v: Vec<_> = s("Mary had a little lamb").split(' ').collect();
assert_eq!(v, vec![s("Mary"), s("had"), s("a"), s("little"), s("lamb")]);

let v: Vec<_> = s("abc1def2ghi").split(|c: char| c.is_digit()).collect();
assert_eq!(v, vec![s("abc"), s("def"), s("ghi")]);

let v: Vec<_> = s("lionXXtigerXleopard").split('X').collect();
assert_eq!(v, vec![s("lion"), s(""), s("tiger"), s("leopard")]);

let v: Vec<_> = s("").split('X').collect();
assert_eq!(v, vec![s("")]);

fn splitn<Sep: CharEq>(&self, count: uint, sep: Sep) -> CharSplitsN<Sep>

An iterator over substrings of self, separated by characters matched by sep, restricted to splitting at most count times.

Example

let v: Vec<_> = s("Mary had a little lambda").splitn(2, ' ').collect();
assert_eq!(v, vec![s("Mary"), s("had"), s("a little lambda")]);

let v: Vec<_> = s("abc1def2ghi").splitn(1, |c: char| c.is_digit()).collect();
assert_eq!(v, vec![s("abc"), s("def2ghi")]);

let v: Vec<_> = s("lionXXtigerXleopard").splitn(2, 'X').collect();
assert_eq!(v, vec![s("lion"), s(""), s("tigerXleopard")]);

let v: Vec<_> = s("abcXdef").splitn(0, 'X').collect();
assert_eq!(v, vec![s("abcXdef")]);

let v: Vec<_> = s("").splitn(1, 'X').collect();
assert_eq!(v, vec![s("")]);

fn split_terminator<Sep: CharEq>(&self, sep: Sep) -> CharSplits<Sep>

An iterator over substrings of self, separated by characters matched by sep.

Equivalent to split, except that the trailing substring is skipped if empty (terminator semantics).

Example

let v: Vec<_> = s("A.B.").split_terminator('.').collect();
assert_eq!(v, vec![s("A"), s("B")]);

let v: Vec<_> = s("A..B..").split_terminator('.').collect();
assert_eq!(v, vec![s("A"), s(""), s("B"), s("")]);

let v: Vec<_> = s("Mary had a little lamb").split(' ').rev().collect();
assert_eq!(v, vec![s("lamb"), s("little"), s("a"), s("had"), s("Mary")]);

let v: Vec<_> = s("abc1def2ghi").split(|c: char| c.is_digit()).rev().collect();
assert_eq!(v, vec![s("ghi"), s("def"), s("abc")]);

let v: Vec<_> = s("lionXXtigerXleopard").split('X').rev().collect();
assert_eq!(v, vec![s("leopard"), s("tiger"), s(""), s("lion")]);

fn rsplitn<Sep: CharEq>(&self, count: uint, sep: Sep) -> CharSplitsN<Sep>

An iterator over substrings of self, separated by characters matched by sep, starting from the end of the string. Restricted to splitting at most count times.

Example

let v: Vec<_> = s("Mary had a little lamb").rsplitn(2, ' ').collect();
assert_eq!(v, vec![s("lamb"), s("little"), s("Mary had a")]);

let v: Vec<_> = s("abc1def2ghi").rsplitn(1, |c: char| c.is_digit()).collect();
assert_eq!(v, vec![s("ghi"), s("abc1def")]);

let v: Vec<_> = s("lionXXtigerXleopard").rsplitn(2, 'X').collect();
assert_eq!(v, vec![s("leopard"), s("tiger"), s("lionX")]);

fn match_indices<'a>(&'a self, sep: &'a Str) -> MatchIndices<'a>

An iterator over the start and end indices of the disjoint matches of sep within self.

That is, each returned value (start, end) satisfies self.slice(start, end) == sep. For matches of sep within self that overlap, only the indices corresponding to the first match are returned.

Example

let v: Vec<(uint, uint)> = s("abcXXXabcYYYabc").match_indices(s("abc")).collect();
assert_eq!(v, vec![(0,3), (6,9), (12,15)]);

let v: Vec<(uint, uint)> = s("1abcabc2").match_indices(s("abc")).collect();
assert_eq!(v, vec![(1,4), (4,7)]);

let v: Vec<(uint, uint)> = s("ababa").match_indices(s("aba")).collect();
assert_eq!(v, vec![(0, 3)]); // only the first `aba`

fn split_str<'a>(&'a self, sep: &'a Str) -> StrSplits<'a>

An iterator over the substrings of self separated by sep.

Example

let v: Vec<_> = s("abcXXXabcYYYabc").split_str(s("abc")).collect();
assert_eq!(v, vec![s(""), s("XXX"), s("YYY"), s("")]);

let v: Vec<_> = s("1abcabc2").split_str(s("abc")).collect();
assert_eq!(v, vec![s("1"), s(""), s("2")]);

fn lines(&self) -> CharSplits<char>

An iterator over the lines of a string (subsequences separated by \n). This does not include the empty string after a trailing \n.

Example

let four_lines = s("foo\nbar\n\nbaz\n");
let v: Vec<_> = four_lines.lines().collect();
assert_eq!(v, vec![s("foo"), s("bar"), s(""), s("baz")]);

fn lines_any(&self) -> AnyLines

An iterator over the lines of a string, separated by either \n or \r\n. As with .lines(), this does not include an empty trailing line.

Example

let four_lines = s("foo\r\nbar\n\r\nbaz\n");
let v: Vec<_> = four_lines.lines_any().collect();
assert_eq!(v, vec![s("foo"), s("bar"), s(""), s("baz")]);

fn char_len(&self) -> uint

Returns the number of Unicode code points (char) that a string holds.

This does not perform any normalization, and is O(n), since UTF-8 is a variable width encoding of code points.

Warning: The number of code points in a string does not directly correspond to the number of visible characters or width of the visible text due to composing characters, and double- and zero-width ones.

See also .len() for the byte length.

Example

// composed forms of `ö` and `é`
let c = s("Löwe 老虎 Léopard"); // German, Simplified Chinese, French
// decomposed forms of `ö` and `é`
let d = s("Lo\u0308we 老虎 Le\u0301opard");

assert_eq!(c.char_len(), 15);
assert_eq!(d.char_len(), 17);

assert_eq!(c.len(), 21);
assert_eq!(d.len(), 23);

// the two strings *look* the same
println!("{}", c);
println!("{}", d);

fn slice(&self, begin: uint, end: uint) -> &Str

Returns a slice of the given string from the byte range [begin..end).

This operation is O(1).

Panics when begin and end do not point to valid characters or point beyond the last character of the string.

See also slice_to and slice_from for slicing prefixes and suffixes of strings, and slice_chars for slicing based on code point counts.

Example

let st = s("Löwe 老虎 Léopard");
assert_eq!(st.slice(0, 1), s("L"));

assert_eq!(st.slice(1, 9), s("öwe 老"));

// these will panic:
// byte 2 lies within `ö`:
// s.slice(2, 3);

// byte 8 lies within `老`
// s.slice(1, 8);

// byte 100 is outside the string
// s.slice(3, 100);

fn slice_from(&self, begin: uint) -> &Str

Returns a slice of the string from begin to its end.

Equivalent to self.slice(begin, self.len()).

Panics when begin does not point to a valid character, or is out of bounds.

See also slice, slice_to and slice_chars.

fn slice_to(&self, end: uint) -> &Str

Returns a slice of the string from the beginning to byte end.

Equivalent to self.slice(0, end).

Panics when end does not point to a valid character, or is out of bounds.

See also slice, slice_from and slice_chars.

fn slice_chars(&self, begin: uint, end: uint) -> &Str

Returns a slice of the string from the character range [begin..end).

That is, start at the begin-th code point of the string and continue to the end-th code point. This does not detect or handle edge cases such as leaving a combining character as the first code point of the string.

Due to the design of UTF-8, this operation is O(end). See slice, slice_to and slice_from for O(1) variants that use byte indices rather than code point indices.

Panics if begin > end or the either begin or end are beyond the last character of the string.

Example

let st = s("Löwe 老虎 Léopard");
assert_eq!(st.slice_chars(0, 4), s("Löwe"));
assert_eq!(st.slice_chars(5, 7), s("老虎"));

fn starts_with(&self, needle: &Str) -> bool

Returns true if needle is a prefix of the string.

Example

assert!(s("banana").starts_with(s("ba")));

fn ends_with(&self, needle: &Str) -> bool

Returns true if needle is a suffix of the string.

Example

assert!(s("banana").ends_with(s("nana")));

fn trim_chars<C: CharEq>(&self, to_trim: C) -> &Str

Returns a string with characters that match to_trim removed from the left and the right.

Arguments

to_trim - a character matcher

Example

assert_eq!(s("11foo1bar11").trim_chars('1'), s("foo1bar"))
let x: &[_] = &['1', '2'];
assert_eq!(s("12foo1bar12").trim_chars(x), s("foo1bar"))
assert_eq!(s("123foo1bar123").trim_chars(|c: char| c.is_digit()), s("foo1bar"))

fn trim_left_chars<C: CharEq>(&self, to_trim: C) -> &Str

Returns a string with leading chars_to_trim removed.

Arguments

to_trim - a character matcher

Example

assert_eq!(s("11foo1bar11").trim_left_chars('1'), s("foo1bar11"))
let x: &[_] = &['1', '2'];
assert_eq!(s("12foo1bar12").trim_left_chars(x), s("foo1bar12"))
assert_eq!(s("123foo1bar123").trim_left_chars(|c: char| c.is_digit()), s("foo1bar123"))

fn trim_right_chars<C: CharEq>(&self, to_trim: C) -> &Str

Returns a string with trailing chars_to_trim removed.

Arguments

to_trim - a character matcher

Example

assert_eq!(s("11foo1bar11").trim_right_chars('1'), s("11foo1bar"))
let x: &[_] = &['1', '2'];
assert_eq!(s("12foo1bar12").trim_right_chars(x), s("12foo1bar"))
assert_eq!(s("123foo1bar123").trim_right_chars(|c: char| c.is_digit()), s("123foo1bar"))

fn is_char_boundary(&self, index: uint) -> bool

Check that index-th byte lies at the start and/or end of a UTF-8 code point sequence.

The start and end of the string (when index == self.len()) are considered to be boundaries.

Panics if index is greater than self.len().

Example

let st = s("Löwe 老虎 Léopard");
assert!(st.is_char_boundary(0));
// start of `老`
assert!(st.is_char_boundary(6));
assert!(st.is_char_boundary(st.len()));

// second byte of `ö`
assert!(!st.is_char_boundary(2));

// third byte of `老`
assert!(!st.is_char_boundary(8));

fn char_range_at(&self, i: uint) -> CharRange

Pluck a character out of a string and return the index of the next character.

This function can be used to iterate over the Unicode characters of a string.

Example

This example manually iterates through the characters of a string; this should normally be done by .chars() or .char_indices.

use dstr::CharRange;

let st = s("中华Việt Nam");
let mut i = 0u;
while i < st.len() {
    let CharRange {ch, next} = st.char_range_at(i);
    println!("{}: {}", i, ch);
    i = next;
}

Output

0: 中
3: 华
6: V
7: i
8: ệ
11: t
12:
13: N
14: a
15: m

Arguments

s - The string

i - The byte offset of the char to extract

Return value

A record {ch: char, next: uint} containing the char value and the byte index of the next Unicode character.

Panics

If i is greater than or equal to the length of the string. If i is not the index of the beginning of a valid UTF-8 character.

fn char_range_at_reverse(&self, start: uint) -> CharRange

Given a byte position and a str, return the previous char and its position.

This function can be used to iterate over a Unicode string in reverse.

Returns 0 for next index if called on start index 0.

Panics

If i is greater than the length of the string. If i is not an index following a valid UTF-8 character.

fn char_at(&self, i: uint) -> char

Plucks the character starting at the ith byte of a string.

Example

let st = s("abπc");
assert_eq!(st.char_at(1), 'b');
assert_eq!(st.char_at(2), 'π');
assert_eq!(st.char_at(4), 'c');

Panics

If i is greater than or equal to the length of the string. If i is not the index of the beginning of a valid UTF-8 character.

fn char_at_reverse(&self, i: uint) -> char

Plucks the character ending at the ith byte of a string.

Panics

If i is greater than the length of the string. If i is not an index following a valid UTF-8 character.

fn as_bytes(&self) -> &[u8]

Work with the byte buffer of a string as a byte slice.

Example

assert_eq!(s("bors").as_bytes(), b"bors");

fn find<C: CharEq>(&self, search: C) -> Option<uint>

Returns the byte index of the first character of self that matches search.

Return value

Some containing the byte index of the last matching character or None if there is no match

Example

let s = s("Löwe 老虎 Léopard");

assert_eq!(s.find('L'), Some(0));
assert_eq!(s.find('é'), Some(14));

// the first space
assert_eq!(s.find(|c: char| c.is_whitespace()), Some(5));

// neither are found
let x: &[_] = &['1', '2'];
assert_eq!(s.find(x), None);

fn rfind<C: CharEq>(&self, search: C) -> Option<uint>

Returns the byte index of the last character of self that matches search.

Return value

Some containing the byte index of the last matching character or None if there is no match.

Example

let s = s("Löwe 老虎 Léopard");

assert_eq!(s.rfind('L'), Some(13));
assert_eq!(s.rfind('é'), Some(14));

// the second space
assert_eq!(s.rfind(|c: char| c.is_whitespace()), Some(12));

// searches for an occurrence of either `1` or `2`, but neither are found
let x: &[_] = &['1', '2'];
assert_eq!(s.rfind(x), None);

fn find_str(&self, needle: &Str) -> Option<uint>

Returns the byte index of the first matching substring

Arguments

needle - The string to search for

Return value

Some containing the byte index of the first matching substring or None if there is no match.

Example

let st = s("Löwe 老虎 Léopard");

assert_eq!(st.find_str(s("老虎 L")), Some(6));
assert_eq!(st.find_str(s("muffin man")), None);

fn slice_shift_char(&self) -> (Option<char>, &Str)

Retrieves the first character from a string slice and returns it. This does not allocate a new string; instead, it returns a slice that point one character beyond the character that was shifted. If the string does not contain any characters, a tuple of None and an empty string is returned instead.

Example

let st = s("Löwe 老虎 Léopard");
let (c, s1) = st.slice_shift_char();
assert_eq!(c, Some('L'));
assert_eq!(s1, s("öwe 老虎 Léopard"));

let (c, s2) = s1.slice_shift_char();
assert_eq!(c, Some('ö'));
assert_eq!(s2, s("we 老虎 Léopard"));

fn subslice_offset(&self, inner: &Str) -> uint

Returns the byte offset of an inner slice relative to an enclosing outer slice.

Panics if inner is not a direct slice contained within self.

Example

let string = s("a\nb\nc");
let lines: Vec<_> = string.lines().collect();
let lines = lines.as_slice();

assert!(string.subslice_offset(lines[0]) == 0); // &"a"
assert!(string.subslice_offset(lines[1]) == 2); // &"b"
assert!(string.subslice_offset(lines[2]) == 4); // &"c"

fn as_ptr(&self) -> *const u8

Return an unsafe pointer to the strings buffer.

The caller must ensure that the string outlives this pointer, and that it is not reallocated (e.g. by pushing to the string).

fn utf16_units(&self) -> Utf16CodeUnits

Return an iterator of u16 over the string encoded as UTF-16.

fn len(&self) -> uint

Return the number of bytes in this string

Example

assert_eq!(s("foo").len(), 3);
assert_eq!(s("ƒoo").len(), 4);

fn is_empty(&self) -> bool

Returns true if this slice contains no bytes

Example

assert!(s("").is_empty());